d problem

+++ task-day7/d.cpp
@@ -2,11 +2,11 @@
 using namespace std;
 int main()
 {
-    int n;
+    long long n;
     cin >> n;
     if (n % 2 == 0)
-        cout << n/2 << endl;
+        cout << n / 2 << endl;
     else
-        cout << -((n+1) /2)  << endl;
+        cout << -((n + 1) / 2) << endl;
     return 0;
 }
